unit-tests/conftest.py
import pytest
import asyncio
import sys
import os

# Add src to path for all tests
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'src'))

# Configure pytest-asyncio
pytestmark = pytest.mark.asyncio